Rename hu.elte.txtuml.diagnostics.animation to *.papyrus
