+++ dev/plugins/hu.elte.txtuml.diagnostics/src/hu/elte/txtuml/diagnostics/animation/papyrus/Animator.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.diagnostics.animation;
+package hu.elte.txtuml.diagnostics.animation.papyrus;
 
 import java.util.HashMap;
 import java.util.Map;
