+++ dev/plugins/hu.elte.txtuml.diagnostics/src/hu/elte/txtuml/diagnostics/session/DiagnosticsPlugin.java
@@ -12,7 +12,7 @@ import java.net.Socket;
 import hu.elte.txtuml.api.model.execution.diagnostics.protocol.Message;
 import hu.elte.txtuml.api.model.execution.diagnostics.protocol.MessageType;
 import hu.elte.txtuml.api.model.execution.diagnostics.protocol.ModelEvent;
-import hu.elte.txtuml.diagnostics.animation.Animator;
+import hu.elte.txtuml.diagnostics.animation.papyrus.Animator;
 import hu.elte.txtuml.utils.Logger;
 
 /**
